Fix typo in definition for "explicit exit"
In the [Daemon Threads section](https://github.com/cs1302uga/cs1302-tutorials/blob/master/threads/brief-intro-threads.md#daemon-threads), the definition provided for the implicit exit condition was incorrectly stated as, "all the threads currently running are *daemon* threads." The table, subsequent discussion, and related material in the videos assumed the correct definition (i.e., all the threads currently running are *daemon* threads). The definition is now corrected, and a footnote is provided with alternative phrasing that aligns with the API documentation provided by the `Thread` class.